requirements.txt
acme>=2.8.0,<3.0.0
josepy>=1.13.0,<2.0.0
numpy>=1.25
pandas>=2.3
ruamel.yaml>=0.17
pytest-asyncio>=0.23
voluptuous==0.13.1
